Did not parallelize, other modifications
